fixed componets in strapi
